Factored threaded backgrounding code out of web.py into background.py
